pyproject.toml
[project]
name = "finance-tracking"
version = "0.1.0"
description = "local python finance dashboard"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dash>=4.0.0",
    "python-dotenv>=1.2.2",
    "sqlcipher3>=0.6.2",
]
